requirements.txt
ipykernel
langchain-openai
langchain
streamlit
sentence-transformers
python-dotenv